+++ modules/@apostrophecms/piece-type/ui/apos/components/AposDocsManager.vue
@@ -146,7 +146,9 @@ export default {
       totalPages: 1,
       currentPage: 1,
       filterValues: {},
-      queryExtras: {},
+      queryExtras: {
+        viewContext: this.relationshipField ? 'relationship' : 'manage'
+      },
       holdQueries: false,
       filterChoices: {},
       allPiecesSelection: {